NTB: Briarwood workarounds
There are a number of issues on the Briarwood platform that require a
workaround on A0 hardware for NTB to work.  A screaming USB interrupt must be
disabled.  Also, one or more BIOs builds do not properly setup the NTB XLAT and
BAR values, thus causing a crash when the first packet is sent across the wire.
The driver must detect and workaround this issue.

These should only be necessary on A0 hardware, and shouldn't be needed for
external releases.  Thus this patch is beign excluded from my NTB kernel
submission.

Signed-off-by: Jon Mason <jon.mason@intel.com>