arm64: Temporary patch for s5p6818 in tlb flush entry
This patch is temporary workaround code for s5p6818 tlb entry flush
routine. This issue occurs when smp on and multi process runs in 32bit
mode. Afterwards resolving, this commit must be reverted.

Since kernel-4.4 has been updated, we need to resurrect this
patch in temporal until we address this issue.

Change-Id: Ie9f6f49a62e805f90eea5ba23c2a59a9009591ef
Signed-off-by: Sungwoo Park <swpark@nexell.co.kr>
Signed-off-by: Chanho Park <chanho61.park@samsung.com>