HACK: arm64: fake armv7l machine for fedora dnf
The dnf command of fedora tries to find its architecture
info from /proc/cpuinfo. This patch is based on Zeng's
patch[1] but it's not sufficient to fake fedora dnf
command because it also read a machine name of uname.
The machine name can't be changed on runtime because it
is decided during initcall. This hack is not good because
it has a potential problem when mixing the binaries between
aarch64 and aarch32[2].
This patch should be reverted if fedora dnf supports aarch32
but it seems not easy[3].

Potential problem:
- Cpuinfo will be 'armv7l' when a aarch32 binary try to read
cpuinfo even though it's on the aarch64 platform.

[1]: https://lkml.org/lkml/2016/4/24/235
[2]: https://lkml.org/lkml/2016/4/25/152
[3]:
https://lists.fedoraproject.org/archives/list/arm@lists.fedoraproject.org/thread/D7MLYMDVNNXKRY5X6JLRYK7VL2XHOLUN/

Change-Id: I36fb8a7b777f59ffbb874a1d4700009fad8b89a4
Signed-off-by: Chanho Park <chanho61.park@samsung.com>