bcmdhd: disabling mixed mode like artik5 or artik10
This patch disables mixed mode like artik5 or 10 by default.
Authentication failure happens when hostapd run with wpa/wpa2 mixed
mode. It is because of wpa/wpa2 mixed for softap.
Currently Artik7 does not support wpa mixed mode, which is the same as
Artik5 or 10. If hostapd sets mixed mode to our driver, it first checks
wpa2 and if it is, it only supports wpa2.

Change-Id: I69ec5fdda3e3136a7e56b7305263e57d7ea43a75
Signed-off-by: Soon-Yong Cho <soonyong.cho@samsung.com>