media: s5k4ecga: fix s_duration function
This patch corrects sequence of s_duration function.
If we set frame duration directly in s_duration function,
i2c write error will occured because power of sensor
will not be turned on at the moment.
Therefore, we have to save frame rate value on state,
then write it to sensor on s_stream function.

Change-Id: I9ab4a6ad7011a5469f6f9585f9ebeb93a6c8a2af
Signed-off-by: Sooman Jeong <sm5.jeong@samsung.com>