drm/panel: s6e8fa0: force to use MIPI_DSI_DCS_SHORT_WRITE_PARAM
This patch is a temporal patch to address the panel is
not turned on correctly when wake up from power kery.
It seems the MIPI_DSI_DCS_SHORT_WRITE_PARAM type message
was not sent at that time. To resolve the issue, we have
to send the gamma update MIPI commands through LONG_WRITE
format.

Change-Id: Ia13070756e94fe4281b2c71fb6a06790e30e5a95
Signed-off-by: Chanho Park <chanho61.park@samsung.com>