media: nexell: decimator: fix operation bug
This patch fix operates bug that appear when decimator test.
It adds linking code between the clippers and decimated.
When you call set_format function of the decimator,
and also modified to call set_fmt function of 'remote subdev.
It was also modified to be terminated without an error
when decimator execution exit('stream off').

Change-Id: Iaae6b2abba978ad0fe86311f4e66f0bf2484610f
Signed-off-by: JongKeun Choi <jkchoi@nexell.co.kr>