net: wireless: sd8977: wlan: enable WoWLAN
Marvell's chipset need to register WoWLAN(Wake on Wireless LAN) trigger in
the wpa_supplicant configuration file because it is not the default. If
host enters sleep without WoWLAN trigger, it makes all wifi device
disconnect with currently connected APs.
This patch is provided by mavell officially. And it enables WoWLAN in the
driver.

Change-Id: Idccf9a8b16352343ef0b8911bb3045b6017a4832
Signed-off-by: Soon-Yong Cho <soonyong.cho@samsung.com>