drm/nexell: correct kzalloc allocation
This patch corrects use of kzalloc allocation. The devm_xxx
function is useful when probe/remove function becuase it can
be manageable by devm_res framework. Except the case, we have
to use kzalloc function. The return value of kzalloc series
function is NULL or correct pointer. Thus, we have to check
whether it is NULL or not.

Change-Id: I12606ae8c38f1d7462e5ed0b8baf941cb3bdd62b
Signed-off-by: Chanho Park <chanho61.park@samsung.com>