net: wireless: bcmdhd: changes DTIM listen interval in suspend mode
A delivery traffic indication map is a kind of traffic indication map
(TIM) which informs the clients about the presence of buffered
multicast/broadcast data on the access point. It is generated within the
periodic beacon at a frequency specified by the DTIM Interval. Beacons
are packets sent by an access point to synchronize a wireless network.
Normal TIMs that are present in every beacon are for signaling the
presence of buffered unicast data. After a DTIM, the access point will
send the multicast/broadcast data on the channel following the normal
channel access rules (CSMA/CA). This helps to have minimum collision and
in effect, increased throughput. In cases where there is not much
interference, or where the number of clients is limited, the DTIM
interval has little or no significance. Usually a value of 1 or 2.

This patch changes DTIM listen interval in suspend mode to 1 from 3. 1
means follow AP's DTIM period. It can receive beacon frame more
frequently.

Change-Id: I06719722fece48f34b1c9cd97cbfb3a9214b7f51
Signed-off-by: Soon-Yong Cho <soonyong.cho@samsung.com>