media: nexell: disable VPU power saving
This patch disables power saving feature of nexell VPU.
When this feature was enabled, we found device goes to completely
freeze state when we repeat opening and closing VPU device, or repeat
rebooting by calling 'reboot' command.
Since this freeze issue is critical, we have to disable VPU power
saving feature until we find root cause and resolution of this issue.

Change-Id: Ib5680282b4a27c00017c005a0b20f79c37a67de6
Signed-off-by: Sooman Jeong <sm5.jeong@samsung.com>