soc: nexell: nxstream: Implement functions about chain
This patch implements functions about nexell stream chaining.
Added functions are following.
- nxs_function_config()
This function is for setting each nxs_dev like format, crop, etc.
- nxs_function_connect()
This function is for tid setting.
- nxs_function_ready()
This function is for setting dirty.
- nxs_function_start()
This function is for starting each nxs_dev.
- nxs_function_stop()
This function is for stopping each nxs_dev.
- nxs_function_find()
This function returns nxs_dev by function identifier.

Change-Id: Iffcb7cbc1b74f79cc7854fe151720f3298dee7df
Signed-off-by: Sungwoo Park <swpark@nexell.co.kr>