soc: nexell: nxs_dev: modify common interrupt value and function
This patch removes the enum "NXS_EVENT_UPDATE" in nxs_event_type
and handle the interrupt as "NXS_EVENT_DONE" interrupt
and how to handle the interrupt is depends on each driver.
and also change the common functions to give nxs_event_type enum
as a argument and change the return value from bool to 32 for
checking each interrupt bit.

Change-Id: I03285ef24f719801fd9216ce2502a278869a349b
Signed-off-by: Hyejungkwon <cjscld15@nexell.co.kr>