soc: nexell: nxs_dmaw: removes checking a open_count part
This patch apply the change of nxs_dev to this driver.
so it use a common function to check, increase and decrease
open_count instead of using atomic function directly.
and fill in a common dump_register function
and change the dev_name to get a clock from dmaw
to dmaw +'the number of the device'.

Change-Id: I33dd046f102d2e993b13e80f05bca36f03ce6a55
Signed-off-by: Hyejungkwon <cjscld15@nexell.co.kr>